==> 06_prototype_intent.ipynb <==
from transformers import AutoTokenizer, AutoModelForSequenceClassification
import torch
import json

#MODEL_DIR = "snips_intent_model_full"

MODEL_DIR = "snips_intent_model_fewshot/snips_intent_model_100" #Can change to 10, 20, 50, 70 or 100 model.

# Load tokenizer & model
tokenizer = AutoTokenizer.from_pretrained(MODEL_DIR)
model = AutoModelForSequenceClassification.from_pretrained(MODEL_DIR)
model.eval()

# Load label2id/id2label
with open("label2id.json") as f:
    label2id = json.load(f)
with open("id2label.json") as f:
    id2label = json.load(f)

# Convert keys to int
id2label = {int(k): v for k, v in id2label.items()}

def predict_intent(text):
    inputs = tokenizer(
        text,
        return_tensors="pt",
        padding=True,
        truncation=True,
        max_length=64
    )

    with torch.no_grad():
        outputs = model(**inputs)
        pred_id = torch.argmax(outputs.logits, dim=1).item()

    return id2label[pred_id]

text = input("Enter query: ")

pred = predict_intent(text)
print("\nPredicted Intent:", pred)
